Force float32 in spotify.go artist.go

+++ spotify.go
@@ -115,9 +115,9 @@ type Followers struct {
 // Image identifies an image associated with an item.
 type Image struct {
 	// The image height, in pixels.
-	Height int `json:"height"`
+	Height float32 `json:"height"`
 	// The image width, in pixels.
-	Width int `json:"width"`
+	Width float32 `json:"width"`
 	// The source URL of the image.
 	URL string `json:"url"`
 }
@@ -222,7 +222,7 @@ func (c *Client) execute(req *http.Request, result interface{}, needsStatus ...i
 				// If the context is cancelled, return the original error
 			case <-time.After(retryDuration(resp)):
 				continue
-                        }
+			}
 		}
 		if resp.StatusCode == http.StatusNoContent {
 			return nil
